SZCR Verification code/szcryzm/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
pytesseract==0.3.10
scikit-learn==1.3.2
scipy==1.11.4

